security domain, update login system

+++ chatapi/src/main/java/sk/avo/chatapi/presentation/users/dto/ResendEmail.java
@@ -3,7 +3,7 @@ package sk.avo.chatapi.presentation.users.dto;
 
 public class ResendEmail extends BaseUser{
 
-    public ResendEmail(Integer id, String username, Boolean isVerified) {
+    public ResendEmail(Long id, String username, Boolean isVerified) {
         super(id, username, isVerified);
     }
 }
